Event_Governance_and_Store/requirements.txt
fastapi
uvicorn
boto3
pydantic
requests 
python-dotenv